CharsLookup

Description

The functions included in this module is mainly for converting char and string into matrix representation that correspond to the pixel in the matrix.

Documentation

lightOn :: Bool

The bit value that would indicate the LED is on or off. The actual value is depended on the polarity of the LED and how it was wired.

lightOff :: Bool

stringToMatrix :: String -> [[Bool]]

Convert a string to the matrix representation.

concatMatrixWithSeparator :: [[Bool]] -> [[Bool]] -> [[Bool]]

Concat two nested lists into one nested list

trimZeros :: [String] -> [String]

Trim the extra columns of zeros

toBoolGrid :: [String] -> [[Bool]]

Convert a nested list of Int of 0s and 1s into Bool where 0 to True and 1 to False.

toIntGrid :: [[Bool]] -> [String]

Convert a nested list of Bool to Int of 0s and 1s where True to 0 and False to 1.

letterToMatrix :: Char -> [[Bool]]

Convert a char to the matrix representation in Bool format

charToStrings :: Char -> [String]

Convert a char to the matrix representation in the string format.

specialStrings :: String -> Maybe [[Bool]]

A set of special strings that would map to different text

rawCharToStrings :: Char -> [String]

The char lookup table